+++ packages/terra-clinical-data-grid/src/subcomponents/Cell.jsx
@@ -74,19 +74,6 @@ class Cell extends React.Component {
     this.getCellStyles = memoize(this.getCellStyles);
   }
 
-  /* eslint-disable class-methods-use-this */
-
-  /**
-   * This function is memoized in the constructor so that for a given width value, the same object reference will be returned.
-   * This allows for repeat renders of the Cell to occur more efficiently if the width value has not changed between renders.
-   */
-  getCellStyles(width) {
-    return {
-      width,
-    };
-  }
-  /* eslint-enable class-methods-use-this */
-
   handleKeyDown(event) {
     if (event.nativeEvent.keyCode === KeyCode.KEY_RETURN || event.nativeEvent.keyCode === KeyCode.KEY_SPACE) {
       const { onSelect } = this.props;
@@ -106,6 +93,18 @@ class Cell extends React.Component {
     }
   }
 
+  /* eslint-disable class-methods-use-this */
+  /**
+   * This function is memoized in the constructor so that for a given width value, the same object reference will be returned.
+   * This allows for repeat renders of the Cell to occur more efficiently if the width value has not changed between renders.
+   */
+  getCellStyles(width) {
+    return {
+      width,
+    };
+  }
+  /* eslint-enable class-methods-use-this */
+
   render() {
     const {
       sectionId,
@@ -123,8 +122,10 @@ class Cell extends React.Component {
       ...customProps
     } = this.props;
 
-    /* eslint-disable react/forbid-dom-props */
+    /* eslint-disable react/forbid-dom-props, jsx-a11y/no-static-element-interactions */
     const theme = this.context;
+    const role = isSelectable ? 'button' : undefined;
+    const tabIndex = isSelectable ? '0' : undefined;
     return (
       <div
         {...customProps}
@@ -133,13 +134,13 @@ class Cell extends React.Component {
         aria-selected={isSelected ? true : undefined}
       >
         <div
-          role="button"
+          role={role}
           className={cx(['content', { selectable: isSelectable, selected: isSelected }])}
           onClick={isSelectable ? this.handleTargetClick : undefined}
           onKeyDown={isSelectable ? this.handleKeyDown : undefined}
           onMouseEnter={onHoverStart}
           onMouseLeave={onHoverEnd}
-          tabIndex={isSelectable ? '0' : undefined}
+          tabIndex={tabIndex}
           ref={selectableRefCallback}
           aria-label={ariaLabel}
         >
@@ -147,7 +148,7 @@ class Cell extends React.Component {
         </div>
       </div>
     );
-    /* eslint-enable react/forbid-dom-props */
+    /* eslint-enable react/forbid-dom-props, jsx-a11y/no-static-element-interactions */
   }
 }