+++ packages/terra-clinical-detail-view/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 3.24.0 - (November 10, 2020)
+
+* Changed
+  * Update wdio tests to remove deprecated Terra.it helpers
+
 ## 3.23.0 - (August 25, 2020)
 
 * Changed